refactor: rename database-related files and update import paths for consistency
